Overhaul the CI setup
Previously, the setup advertised that it was building with various
different GHC versions, but... it wasn't. It was always building with
whichever version of GHC was specified by the Stackage snapshot.

Now, there are two separate build matrices. One builds with Stack
according to the Stackage snapshot specified in stack.yaml. The other
builds with cabal-install, and does so across several different versions
of Cabal and GHC.