Use Stackage LTS 23; constrain the tls package to pre-2.0 versions
This breaks the Cabal CI workflows because Cabal can't find a set of
packages that satisfies the tls < 2.0 bound along with all of the other
version bounds. There obviously *is* such a set of packages--Stack is
able to find one, at least when explicitly told which version of tls to
use--but I'm not sure how to tell Cabal that.

This fixes issue #1.