Support both pre- and post-2.0 versions of the tls library
Version 2.0 of the tls library dropped support by default for the kind
of configuration that is currently being used by the Pinboard API
server. (It only supports up to TLS 1.2, and it doesn't support the EMS
extension.) I tried to work around this by adding a version bound on the
tls package so we would always get a pre-2.0 version. But this is an
unnecessarily blunt solution: newer versions of tls *can* connect to the
Pinboard server, it just requires a little bit of configuration. This
commit introduces this configuration.

See issue #1.